Start playing with infix operator handling
These all need to be converted into message sends of their appropriate
type and name so we can't just implement a simple adder in our
interpreter. Also decided to get rid of all prefix operators (- and !)
for now and see if they should really exist or not.

Interpreter tests for this are commented out, there's a bit more work
that needs to be done before that level of things is working.