Evaluate block literals into block objects
This is the first step to getting these things actually evaluated.
Working through this pointed to a split in the Object modelling, where
we have objects with explicit values (numbers, booleans, strings), and
Objects with other data (blocks), so now providing a ValueObject
interface to help differentiate the two.

This difference is implementation only and will not mean anything to the
system.