Start defining our core object library
Manipulation of core values like numbers and strings needs to happen in
javascript land for now, so we need to set up a library of built-in
functions on our base objects (Number and String for now).

No type-checking these arguments yet. Passing in something other than a
Number will give NaN, probably.