Replace the internal-only environment and scopes with objects!
This applies two concepts to the runtime: the World and Spaces. There's
intended to be one World per run, but in that World are many Spaces. All
code execution takes place in a Space, and Spaces are where locals are
stored and used as contextual scoping. Blocks are then sub-spaces of
their parent space so locals are stored but outside values are still
usable.