Fix the parser to let blocks be used in arguments
This was fun to track down because it required a following expression so
that `while(this.peekToken())` eval'd to true. Parsing a block inside of
a call site led to the next iteration of the while loop to have
currToken set to the close param, which sent the while loop through
again and cause all sorts of havok.