Expose the VM object to abstract "Interpreter"
The interpreter itself is an implementation detail of how code gets
executed and isn't something that should be front-and-center. The VM
will now keep track of the current execution context as well as loading
in the core libraries. This required changes to the tests as more of
them are running code and depending on core library functionality. I
expect some significant changes to the test suite to make this all more
consistent.

Also, holy crap npm dependencies are rediculous. Will probably find a
better way of globbing library files that doesn't load up another 100+
dependencies.