Lex and parse comments
Comments can be attached to AST nodes to keep track of public
documentation for things like constants and the public API of a given
object. Still need to figure out how to get from the AST to the actual
objects. Ideally the doc strings are attributes on the Slot itself,
though we don't have a concept of a Slot yet (it's just a javascript
string). That probably has to happen first.

This work did lead to a decision that if you want to document multiple
messages on an object, you need to add them all together either with
`.new` or via `addSlots`.