Support comments as first class citizens
This takes the comment values from the lexer and parser and actually
applies them to the objects in question. This works on assignment,
`new`, and `addSlots` to provide hooks for people to write documentation
on objects and the messages they support.

Even though this is the fourth or fifth attempt at getting this working
it's still kind of janky and probably not tested enough, however,
`examples/docs.lang` works and shows this feature off decently well.