Truncate super long code blocks in error reporting
It's possible for an error report to span many lines, like with an
unclosed `}`, so lets take steps to keep that from spamming the user
with useless output. Now splits the output into beginning and end
sections if the output would be longer than necessary.

Also made the decision that the code blocks should start 2 spaces in to
match what will be commonly used in descriptions and to differentiate it
from the error message itself.