Move lexing/parsing out of the interpreter
I'm starting into runtime errors and need to give some serious thought
into how I'm going to handle this. It's more than just throwing an error
and quitting, this system needs to be a run-time live-updating VM that
gives people the ability to see errors when they happen. This will
require a bit of thought and design into how this needs to work, and
that should help point me into what runtime errors actually are.