Keep track of lexer tokens in the parsed AST tree
These tokens are necessary for us to back-track into the source to find
out where an error happened and how to build a nice error report. Also
pointed me to some string numbers that should have been actual numbers
in the test, not sure why the test didn't fail on those before.