Initial pass of runtime errors!
This ended up touching a lot of stuff. While SlotNotFoundError was
trivial to implement, NotABlockError showed me that I had left out an
important part of the puzzle: knowing if the code wants to treat a value
as a block or not (`obj.message` had the exact same parse tree as
`obj.message()`).

To fix this I turned `message()` into syntax sugar that gets expanded
into `message.call()`. This then changed the expected AST for a ton of
tests and worked on cleaning that up a bit as well.

From here I had to tackle a second problem: making sure we're
calling the block in the right context. With the double-jump from the
envforced `call()` structure we now need to keep a pointer back to the
originating object and set up context that way.

Which then showed me that my `objectId` builtin now broke this entire
contract. It happened to work properly due to the ambiguity earlier but
with that solved it now caused an infinite loop. I haven't figured out
how to stop the infinite loop when the code is written wrong (that's
next) but now fixed the usage of objectId to be a block call.

And now I have runtime errors and output that look like lexer and
parsing errors... and point to the wrong part of the code. But progress!