Get Closure status and `self` working properly
The `context` idea ended up being a dud mainly because it would cause
double execution of the tree and would lead to really strange execution
problems. It was also a problem because I was trying to keep the `self`
status on the method itself, which isn't possible while supporting what
we need.

I've ended up here wrapping methods in activation records that are used
to keep track of the calling object and possibly other information in
the future. This way if there are chains to, say, `Number.+`, each call
will keep track of the proper owning number.

Also now keeping track of the enclosing scope when we evaluate a block
literal, which allows us to have proper closure status in our blocks.