Better handling of implicit self for block invocation
Having `self` in the AST was proving to cause a couple of problems when
figuring out who owns a block and providing that block with its own
space and the correct value in that space's `self` slot.

With this all tests and situations I've been able to test show that
`self` is right, block ownership is right, and blocks are properly closures.