Array iteration!
This commit gets Array.each in place, evaluating a block for every entry
in the array. The test fails though as it points out that our local
scoping rules are currently following Python-esque rules (always local)
instead of what I want which is Ruby-esque (if one matches, update it
first before creating local).