Use Ruby-style scope lookup rules
With this we will look up the scope parents to find who actually owns a
slot with a given name, and keep using that parent as we change the
value of said slot.

Compare this to Python, where the tests in place would fail because each
nested `a = x` line would create a new local variable `a` that is
unrelated to the outer versions, and how this system was working before.

I find Ruby's version of scoping much more intuitive than Python's. You
can make Python work the Ruby way but it requires a new keyword to
trigger it and that's not something easily explained or understood.