Make use of parents to better implement scope lookup
Instead of a two-step process where we have to look in the Spaces as
well as what's in `self`, just inject the `self` into the parents list,
then it all Just Works (in theory, and the tests pass).

Realized I wasn't doing the second step (self lookup) in all places
leading to weird SlotNotFound errors in deeply nested blocks. This work
fixes those situations.