Implement Object.is to provide runtime type checking
This will return true or false depending on if an object is the same
object or if its parents have the requested object.

Through this realized I needed to make a more generic tool for doing
lookups in object parent trees and did some refactoring around the other
parts that already run this lookup.