Fix bug where ActivationRecord always had a receiver
In what was intended to just be documentation, I inadvertantly created a
painful bug that was causing `self` to be null in the strangest of
places. The interpreter expects the `reciever` slot to not exist when
there isn't a receiver to apply, but my change made sure that it was at
least always the Null object.