Move more built-in functions to live in BuiltIn
I'd like to keep explicitly built-in functions as minimal as possible.
Making explicit use of BuiltIn in a proxy pattern lets the language
document as much of its API directly as possible.

For situations like new() and addSlots(), those functions cannot
currently be implemented in the language due to the var-arg nature so
they have to stay direct at this time.