Switch to TSPM over tspath
tspath does not support shebangs, and when I tried to add that support
it started failing on the .map files as well. Through this I realized
that it was tspath that was outputting compressed JS with no option to
not do that. I found `ef-tspm` instead and it does everything I want.

Ok, not everything. It still gets shebangs wrong. It does comment them
out but then doesn't revert that when writing the final file out.

 /sigh