Provide default-value tracking for initial state of child objects
This feature came out of how I automatically wrote the test framework
and how I expected the system to work, and it slightly blurs the lines
between a true prototype-based language and a class-based one.

You can see an example in the object_spec.lang file below but what I
realized is that objects do inherit everything from their parent on
`new`, including state. I could have gone the route of documenting that
this is just how the system is "supposed" to work and provide an
`init()` slot for resetting such things but to me that breaks principle
of least surprise.

So what I'm doing here is that when a call to `new()` is made, the
system will find all default values (defined by the source code) of
non-block slots and insert those slots and their default values into the
new object.

There isn't currently a way to turn off this behavior and I'm not sure
we should as expecting data sharing between children and parents could
easily lead to code flow that's super hard to track down and follow.