Move from npm scripts to a Makefile
`npm run` is just a PITA trying to combine multiple different commands
in different ways, and having to type all of that all of the time is
frustrating. Time to grab the simplest thing that will work: Makefile.

Everyone has `make`
It's just bash
I don't need to do much