Begin working on a task runner utility
For now this will be a part of the standard library. I want something
like `rake` or `mix` for running things on the command line and building
various tasks that the program needs to run. The first task will be the
`test` task which will then be used to run our spec suite as a single
unit and finally providing a way for Travis to fail if the spec suite
fails!