Pass arguments to task blocks
Right now we will pass one argument down as an array of all arguments
provided on the command line. Want this to change to be a splat so it
can be nothing or more and task blocks can use regular parameters.

Also now exposing javascript's Array.slice to help with this.