Update VM to bubble instead of catch all errors
Now that I've got `load` running, I need to be able to catch exceptions
caused when `load`-ing a file via `try()`. However to do that the Lexer
and Parser both needed to be updated to let their exceptions bubble up.
This also led to moving error reporting out of the VM itself and into
the CLI, such that if you're running the CLI and errors get to the top,
nice reports happen, but otherwise the system is able to catch and
handle anything itself.