Add `into` to load and start adding runtime error tests
Adding the runtime error test showed me that I needed a way to load a
file into a specific context and not to the World as we don't want these
tests to infect the World and cause issues in other tests (it did).

In doing this I realized I didn't like the attempt to split World and
stdlib into multiple spaces so now it's back to just installing
everything into the World then setting a new space for code eval.