Load and run the spec suite in the browser!
Built a quick FakeFS object to let us put "files" in the browser by path
and content. Use this to then consolidate loading for core / stdlib
files and now provides a nice way to add the spec suite where it isn't
loaded automatically. Also providing the full project-dir local paths to
the web to make debugging a little easier to handle.