First pass of a web-safe interpreter
The naiive first interpreter hogs the whole node/web js process and only
accidentally lets it go during things like IO. This is not going to be
web friendly in any sense of the term so I started looking into
alternatives. The main requirement to be web safe is that the interpeter
MUST defer to the runtime at regular intervals to let it handle the
things it needs to run.

This means the interpeter itself needs to be written in an asynchronous
manner, even though execution is still single-threaded and linear,
we can no longer rely on function return values and we definitely can't
use a recursive tree walker anymore.

The pattern I've chosen here is a double-stack setup: a code stack and a
data stack, after seeing a reserach paper about doing this for
low-memory devices [1]. This pattern lives in-between the world of raw
interpretation and a full byte-code compiler and VM. As code comes in,
it gets pushed onto a stack. Data that's needed for each step also gets
pushed onto a data stack. Return for a specific chunk of code is then
handled by a Promise and a specific ReturnValue node type.

I spent some time thinking about using WebWorkers (and the node Worker
API) but realized that this would be too restrictive to my needs and
would require an extensive communication protocol to fully control the
threaded interpreter. While Web Workers are something we should support,
it will be an add-on when the new VM is running well itself.

[1] http://shodhganga.inflibnet.ac.in/handle/10603/43114
    https://www.semanticscholar.org/paper/Optimizing-AST-Node-for-Java-Script-Compiler-A-for-Pattanayak-Patra/21ca570614a5862e79d47883d132e6977d1df0a3