Basic Assignment and Identifier lookup now working
Realized that we needed to try to make the internals of the interpreter
as synchronous as possible and not rely on Promise objects, so the
ReturnValue object and its underlying Promise are both used in the
evaluation process.

Pre-setting the size of the array didn't do anything with push/pop, the
array still got bigger and smaller with push/pop, so that was a waste.