Loading the core and stdlibs, finished the block tests
Due to the async nature of the evaluator, every file ends up loading
through async/await, but we need to not run our tests until all of the
core and stdlib files are fully evaluated. As such, time to set up a new
web-safe version of the VM that handles this.

Also got built-in function calls working again and started converting
all of those to the new format.