Re-implement Array.each via recursion
Trying to get this built-in updated for the web-safe version I realized
that to make this new VM re-entrant would require a significant re-think
of how it's used externally, and I wasn't sure what that would need to
look like. So I tried to implement the looping in-language using the
tool meant for such a thing: recursion! After debugging a *really* weird
issue where code was getting evaluated backwards (turns out
Array.prototype.reverse() modifies the array in-place *and* returns
the reversed array, oops!) it works well!

Now I'll need to set up some tail-call handling on this kind of pattern
but that's for later.