Re-implement `World.load`
This is the first technically re-entrant built-in but I'm able to do it
in a way that doesn't have to wait for the loaded code to finish
working. Instead I rely on the code stack to do the Right Thing for me,
though it does finish the `load`'s block before it starts processing the
requested file's code.