Rebuild argument handling to fill in an object over using the stack
I ran into a situation via the Test framework where complex arguments
(like [] == []) were breaking when it was the first of multiple
arguments. It took a while to track down what was going on but what I
found was that pushing multiple PushArgument nodes in succession was
leading to the data stack getting confused for the later PushArgument
(the last argument ran never failed to store the right value). As
everything else with the data stack works I decided to change things
around and instead of pushing argument results on the stack, I will push
them into the StartBlock node that's awaiting execution. This fits the
pattern of other commands I use and removes a lot of weird object
wrapping and unwrapping to keep argument names and any attached comments
around.

Something broke with exceptions though, need to track that down now.