Fix `load`'s `into` parameter to work properly again
Because built-ins can't be re-entrant and because `load` needs to be so
at this time, the file being loaded doesn't actually start being
evaluated until after `load` itself finishs evaluation. This then
triggers a "popSpace" which puts us in the space of `load`'s caller, and
not in the space that the loaded file is supposed to be evaluated in.

So, to work around this, I'm adding a new command that just sets the
current space to `into` but also properly fills in `load`'s own
previousSpace to the correct value which we don't know until we're
evaluating PushCurrentSpace.

All of this should be removable once there's enough functionality to
implement `load` in-language, but that will require not only more File
code but also exposing the VM itself into the language.