Move to a per-space only local stack
I realized that using the single global data stack was leaving a lot of
data around. Every expression has a value and those values all end up on
the stack regardless if they're used or not. Instead of trying to figure
out a way to "clean up" the stack, I can do that automatically by moving
the stack to always be space-only. Then when the space is cleaned up, so
is the stack.