Update EachParent and FindIn to be be short-circuit-able
This was causing severe slowdowns even when the call stack was only
slightly deep, and really went bad when running the spec benchmark (each
is recursive, times running the test suite 100 times == SUPER deep
stack). Some quick CPU profiling showed that over 50% of the CPU was in
EachParent.

With this short circuiting, the runtime of the spec suite is almost
fully back to what it was before the web-safe work, dropping from ~500ms
down to just over 200ms.