Clean up and get rid of an if block
This is the hot loop. I don't think it makes a huge difference but I
like how this looks better with the while conditional being the trigger
instead of a nested conditional and a `break`.

Also make sure the node is popped *after* the time check. If done the
other way around, you'll pop a node, see that the time window is up, and
lose that popped node entirely. That can lead to really weird errors!